# docs/baseline/openapi.snapshot.json
{
  "components": {
    "schemas": {
...
      "DataResponse_OrgSettingsResponse_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/OrgSettingsResponse"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaResponse",
            "default": {}
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[OrgSettingsResponse]",
        "type": "object"
      },
...
      "DataResponse_dict_": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "title": "Data",
            "type": "object"
          },
          "meta": {
            "$ref": "#/components/schemas/MetaResponse",
            "default": {}
          }
        },
        "required": [
          "data"
        ],
        "title": "DataResponse[dict]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "OrgApiKeyCreateRequest": {
        "properties": {
          "name": {
            "default": "dashboard",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "title": "OrgApiKeyCreateRequest",
        "type": "object"
      },
...
      "OrgMemberCreateRequest": {
        "properties": {
          "email": {
            "maxLength": 320,
            "minLength": 3,
            "title": "Email",
            "type": "string"
          },
          "role": {
            "default": "viewer",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "OrgMemberCreateRequest",
        "type": "object"
      },
      "OrgMemberUpdateRequest": {
        "properties": {
          "role": {
            "description": "admin|engineer|compliance|viewer",
            "title": "Role",
            "type": "string"
          }
        },
        "required": [
          "role"
        ],
        "title": "OrgMemberUpdateRequest",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "The control plane for AI agent networks",
    "title": "Nexra API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/orgs/api-keys": {
      "get": {
        "operationId": "list_api_keys_v1_orgs_api_keys_get",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Api Keys",
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "create_api_key_v1_orgs_api_keys_post",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgApiKeyCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Api Key",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/orgs/api-keys/{key_id}": {
      "delete": {
        "operationId": "revoke_api_key_v1_orgs_api_keys__key_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "key_id",
            "required": true,
            "schema": {
              "title": "Key Id",
              "type": "string"
            }
          },
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revoke Api Key",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/orgs/me": {
      "get": {
        "operationId": "get_org_settings_v1_orgs_me_get",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OrgSettingsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Org Settings",
        "tags": [
          "organizations"
        ]
      },
      "patch": {
        "operationId": "update_org_settings_v1_orgs_me_patch",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgSettingsUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_OrgSettingsResponse_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Org Settings",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/orgs/members": {
      "get": {
        "operationId": "list_members_v1_orgs_members_get",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Members",
        "tags": [
          "organizations"
        ]
      },
      "post": {
        "operationId": "create_member_v1_orgs_members_post",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgMemberCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Member",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/orgs/members/{member_id}": {
      "delete": {
        "operationId": "delete_member_v1_orgs_members__member_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "title": "Member Id",
              "type": "string"
            }
          },
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Member",
        "tags": [
          "organizations"
        ]
      },
      "patch": {
        "operationId": "update_member_v1_orgs_members__member_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "member_id",
            "required": true,
            "schema": {
              "title": "Member Id",
              "type": "string"
            }
          },
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgMemberUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Member",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/orgs/register": {
      "post": {
        "description": "Create a new organization and return its API key.\n\nThe API key is returned exactly once in this response.\nIt is bcrypt-hashed before storage and cannot be retrieved again.",
        "operationId": "create_organization_v1_orgs_register_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OrgCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Organization",
        "tags": [
          "organizations"
        ]
      }
    },
    "/v1/orgs/session": {
      "get": {
        "operationId": "get_org_session_v1_orgs_session_get",
        "parameters": [
          {
            "description": "Bearer nx_live_...",
            "in": "header",
            "name": "authorization",
            "required": true,
            "schema": {
              "description": "Bearer nx_live_...",
              "title": "Authorization",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "X-User-Email",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Email"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataResponse_dict_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Org Session",
        "tags": [
          "organizations"
        ]
      }
    },
...
  }
}